6/14 리뷰 작성 기능

+++ app/src/main/java/com/example/itsapp/model/vo/Device.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class Device(
     @SerializedName("DEVICE_NAME") var deviceName: String,
     @SerializedName("DEVICE_BRAND") var deviceBrand: String,
-    @SerializedName("REVIEW_POINT") var reviewPoint: Double,
+    @SerializedName("REVIEW_POINT") var reviewPoint: Float,
     @SerializedName("REVIEW_COUNT") var reviewCount: Int,
     @SerializedName("DEVICE_PRICE") var devicePrice: String,
     @SerializedName("REVIEW_POINT_5_COUNT") var reviewPoint5Count :Int,
